优化Goupon、Goods、User等服务的创建对象流程

+++ shop-goods-service/src/main/java/com/my/shop/mapper/GoodsLogMapper.java
@@ -21,6 +21,6 @@ public interface GoodsLogMapper {
 
     void add(@Param("goodsLog") GoodsLog goodsLog);
 
-    Integer findMaxId();
+    BigInteger findMaxId();
 
 }
